fix bug2

+++ db/activity-3.0.0-delta-schema.sql
@@ -3,10 +3,10 @@ ALTER TABLE `eh_activity_roster` ADD COLUMN `pay_flag` TINYINT(4) DEFAULT '0' NU
 ALTER TABLE `eh_activity_roster` ADD COLUMN `order_no` BIGINT(20) NULL;
 ALTER TABLE `eh_activity_roster` ADD COLUMN `order_start_time` DATETIME NULL;
 ALTER TABLE `eh_activity_roster` ADD COLUMN `vendor_type` VARCHAR(32) NULL COMMENT '10001: alipay, 10002: wechatpay';
-ALTER TABLE `eh_activity_roster` ADD COLUMN `pay_amount` INT(11) NULL;
+ALTER TABLE `eh_activity_roster` ADD COLUMN `pay_amount` DOUBLE NULL;
 ALTER TABLE `eh_activity_roster` ADD COLUMN `pay_time` DATETIME NULL;
 ALTER TABLE `eh_activity_roster` ADD COLUMN `refund_order_no` BIGINT(20) NULL;
-ALTER TABLE `eh_activity_roster` ADD COLUMN `refund_amount` INT(11) NULL;
+ALTER TABLE `eh_activity_roster` ADD COLUMN `refund_amount` DOUBLE NULL;
 ALTER TABLE `eh_activity_roster` ADD COLUMN `refund_time` DATETIME NULL;
 ALTER TABLE `eh_activity_roster` ADD COLUMN `status` TINYINT(4) DEFAULT '2' NULL COMMENT '0: cancel, 1: reject, 2:normal';
